Organizing

+++ mapFolding/lolaBeDRY.py
@@ -2,19 +2,19 @@ import numpy
 from typing import List, TypedDict, Final, Literal
 
 class countFoldsTask(TypedDict):
-    activeGap1ndex: int
-    activeLeaf1ndex: int
+    activeGap1ndex: numpy.uint8
+    activeLeaf1ndex: numpy.uint8
     connectionGraph: numpy.ndarray
     countDimensionsGapped: Literal[2]
-    dimensionsTotal: int
+    dimensionsTotal: numpy.uint8
     foldsTotal: int
     gapRangeStart: Literal[3]
     leafAbove: Literal[0]
     leafBelow: Literal[1]
-    leavesTotal: int
+    leavesTotal: numpy.uint8
     listDimensions: List[int]
     potentialGaps: numpy.ndarray
-    taskIndex: int
+    taskIndex: numpy.uint8
     track: numpy.ndarray
 
 leafAbove: Final[Literal[0]] = 0
